requirements.txt
sounddevice
numpy
scipy
openai-whisper
ollama
customtkinter
packaging
